+++ src/pages/CRUD.jsx
@@ -1,7 +1,8 @@
-import React from "react";
+import React, {useState}from 'react'
+
 
 function CRUD() {
-  return <div>CRUD</div>;
+     <div>CRUD</div>
 }
 
 export default CRUD;
